Update template builder with changes from f707b4a
