+++ shared_model/builders/protobuf/builder_templates/transaction_template.hpp
@@ -24,6 +24,7 @@
 
 #include "block.pb.h"
 #include "commands.pb.h"
+#include "primitive.pb.h"
 
 #include "amount/amount.hpp"
 #include "builders/protobuf/helpers.hpp"
@@ -95,8 +96,6 @@ namespace shared_model {
       }
 
      public:
-      using ModelType = BT;
-
       TemplateTransactionBuilder(const SV &validator = SV())
           : stateless_validator_(validator) {}
 
@@ -126,7 +125,7 @@ namespace shared_model {
           auto command = proto_command->mutable_add_asset_quantity();
           command->set_account_id(account_id);
           command->set_asset_id(asset_id);
-          addAmount(command->mutable_amount(), amount);
+          initializeProtobufAmount(command->mutable_amount(), amount);
         });
       }
 
@@ -134,8 +133,9 @@ namespace shared_model {
                    const interface::types::PubkeyType &peer_key) const {
         return addCommand([&](auto proto_command) {
           auto command = proto_command->mutable_add_peer();
-          command->set_address(address);
-          command->set_peer_key(crypto::toBinaryString(peer_key));
+          auto peer = command->mutable_peer();
+          peer->set_address(address);
+          peer->set_peer_key(crypto::toBinaryString(peer_key));
         });
       }
 
@@ -290,7 +290,7 @@ namespace shared_model {
           auto command = proto_command->mutable_subtract_asset_quantity();
           command->set_account_id(account_id);
           command->set_asset_id(asset_id);
-          addAmount(command->mutable_amount(), amount);
+          initializeProtobufAmount(command->mutable_amount(), amount);
         });
       }
 
@@ -306,17 +306,11 @@ namespace shared_model {
           command->set_dest_account_id(dest_account_id);
           command->set_asset_id(asset_id);
           command->set_description(description);
-          addAmount(command->mutable_amount(), amount);
+          initializeProtobufAmount(command->mutable_amount(), amount);
         });
       }
 
-      template <typename T>
-      auto addAnyCommand(const T &command) {
-        return addCommand(
-            [&](auto proto_command) { proto_command->CopyFrom(command); });
-      }
-
-      BT build() const {
+      auto build() const {
         static_assert(S == (1 << TOTAL) - 1, "Required fields are not set");
 
         auto answer = stateless_validator_.validate(
